control al eliminar Session y Weekly Routine

+++ backend/src/main/java/com/gym/backend/business/services/SessionService.java
@@ -62,6 +62,12 @@ public class SessionService {
 
     @Transactional
     public void delete(long sessionId) {
+        long count = routineDayRepository.countBySessionId(sessionId);
+        if (count > 0) {
+            throw new IllegalArgumentException(
+                    "No se puede eliminar esta sesión, porque está asociada a rutinas diarias.");
+        }
+
         repository.deleteById(sessionId);
     }
 }
